+++ stories/02-Design System/04-DataEntry/DatePicker/data.ts
@@ -470,6 +470,14 @@ const CommonArgs = {
             },
             category: 'Common API'
         }
+    },
+    readonly: {
+        name: 'readonly',
+        description: 'Boolean to set the input in readonly mode',
+        table: {
+            defaultValue: {summary: false},
+            category: 'Common API'
+        }
     }
 };