+++ stories/02-Design System/04-DataEntry/Input/data.ts
@@ -351,5 +351,13 @@ export const argTypes = {
             },
             category: 'Input'
         }
+    },
+    readonly: {
+        name: 'readonly',
+        description: 'Boolean to set the input in readonly mode',
+        table: {
+            defaultValue: {summary: false},
+            category: 'Common API'
+        }
     }
 };