+++ stories/02-Design System/04-DataEntry/Select/data.ts
@@ -505,6 +505,14 @@ const SelectArgTypes = {
             },
             category: 'Select'
         }
+    },
+    readonly: {
+        name: 'readonly',
+        description: 'Boolean to set the input in readonly mode',
+        table: {
+            defaultValue: {summary: false},
+            category: 'Common API'
+        }
     }
 };